Ändrade några ints till float
ändrade damage och health till floats så att man kan ha en damage mult från att levla upp

+++ Programveckor 2025/Assets/Scripts/Enemy/EnemyUnit.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [RequireComponent(typeof(Rigidbody2D))]
 public class EnemyUnit : UnitBase
 {
-    [SerializeField] int damage;
+    [SerializeField] float damage;
 
     [HideInInspector] public bool spared;
